3.6 update

+++ test_cuda.py
@@ -2,7 +2,7 @@ import torch
 
 def check_device():
     USE_CPU  = True
-    dtype = torch.float64
+    dtype = torch.float
     if USE_CPU and torch.cuda.is_available():
         device = torch.device('cuda')
     else:
